Create Palindrome.cpp
This program prints the list of numbers which are palindrome (same whether you read it forward or backward. Like: "mom" or "tacocat.") in the given range of Numbers. 
This is the contribution for the hacktober fest 2020. Can you please add the tag to it and merge it to your repo.
Thank you